assignment/requirements.txt
beautifulsoup4==4.8.1
Django==2.2.7
django-bootstrap4==1.0.1
django-webpack-loader==0.6.0
djangorestframework==3.10.3
numpy==1.17.4
pandas==0.25.3
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
